requirements.txt
Pillow>=10.0.0
pdf2image>=1.16.3
img2pdf>=0.4.4
ttkthemes>=3.2.2 